Apply static casts to reduce implicit casts 2

+++ tree/candidateset.cpp
@@ -116,7 +116,7 @@ vector<string> CandidateSet::getBestTreeStringsForProcess(int numTree) {
     }
     
     if (numTree == 0 || numTree > alltrees.size()) {
-        numTree = alltrees.size();
+        numTree = static_cast<int>(alltrees.size());
     }
     size_t cnt = 0;
     vector<string> res;
@@ -269,14 +269,14 @@ int CandidateSet::update(string newTree, double newScore) {
     }
     ASSERT(topologies.size() == size());
 
-    treePos = distance(candidateTreeIt, end());
+    treePos = static_cast<int>(distance(candidateTreeIt, end()));
 
     return treePos;
 }
 
 vector<double> CandidateSet::getBestScores(int numBestScore) {
     if (numBestScore == 0) {
-        numBestScore = size();
+        numBestScore = static_cast<int>(size());
     }
     vector<double> res;
     for (reverse_iterator rit = rbegin(); rit != rend() && numBestScore > 0; rit++, numBestScore--) {
@@ -418,7 +418,7 @@ void CandidateSet::removeWorstTree() {
 
 int CandidateSet::computeSplitOccurences(double supportThreshold) {
     candSplits.clear();
-    candSplits.setNumTree(size());
+    candSplits.setNumTree(static_cast<int>(size()));
 
     /* Store all splits in the best trees in candSplits.
      * The variable numTree in SpitInMap is the number of trees, from which the splits are converted.
